Crazy Mod 1.2.1
 EP's original forum post:

Version 1.2.1, unfortunately. 1.2.0 was broken in netplay.

Changes:
The netplay bug is fixed, but at the cost of:
The number of shots in the arena is now limited again, although the limit is around 255 rather than 100.

Also, the Shofixti's regeneration is no longer proportional to the amount of crew, and is generally slower.